digitinspector tests: added more robust cases & done

+++ Lewis/Lewis/Logic/Scrambler.cs
@@ -6,7 +6,7 @@ namespace Lewis.Logic
 {
     public class Scrambler
     {
-        public int Number { get; set; }
+        public long Number { get; set; }
         public HashSet<string> GetAllPermutationStrings()
         {
             var results = new List<string>();
@@ -18,7 +18,7 @@ namespace Lewis.Logic
             return results.ToHashSet();
         }
 
-        private void Permute(char[] digits, int currentIndex, List<string> results)
+        private void Permute(char[] digits, long currentIndex, List<string> results)
         {
             // Base case: if we've reached the end of the array, we have a complete permutation
             // also running this will go outside the array bounds
@@ -29,7 +29,7 @@ namespace Lewis.Logic
             }
 
             // Recursive case: the current index with every index after it
-            for (int i = currentIndex; i < digits.Length; i++)
+            for (long i = currentIndex; i < digits.Length; i++)
             {
                 Swap(digits, currentIndex, i);      // 1. Swap the current element with the loop index
                 Permute(digits, currentIndex + 1, results); // 2. Recurse for the next position
@@ -37,14 +37,14 @@ namespace Lewis.Logic
             }
         }
 
-        private void Swap(char[] array, int a, int b)
+        private void Swap(char[] array, long a, long b)
         {
             char temp = array[a];
             array[a] = array[b];
             array[b] = temp;
         }
 
-        public Scrambler(int number)
+        public Scrambler(long number)
         {
             Number = number;
         }
